# appointment-manager-api/src/main/resources/api-definition.yaml
openapi: 3.0.0
info:
  title: Online Booking Service API
  version: 1.0.0
  description: API for managing online bookings for various companies.
...
paths:
...
  /users/{userId}/appointments:
    get:
      summary: Get list of appointments for a user
      operationId: getAppointmentsByUser
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A list of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
...
  /users:
    post:
      summary: Create a new user
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    get:
      summary: Get user by ID
      operationId: getUserById
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: A single user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    put:
      summary: Update user by ID
      operationId: updateUserById
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    delete:
      summary: Delete user by ID
      operationId: deleteUserById
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted successfully
  /users/{userId}/avatar:
    post:
      summary: Upload user avatar
      operationId: uploadUserAvatar
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Avatar uploaded successfully
        '400':
          description: Bad request, invalid file format
        '404':
          description: User not found
    get:
      summary: Get user avatar
      operationId: getUserAvatar
      tags:
        - Users
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User avatar retrieved successfully
          content:
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: User avatar not found
...
components:
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        userId:
          type: integer
        employeeId:
          type: integer
        serviceId:
          type: integer
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
      required:
        - userId
        - employeeId
        - serviceId
        - startTime
        - endTime
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        phoneNumberConfirmed:
          type: boolean
        hasAvatar:
          type: boolean
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - name
        - password
        - phoneNumber
...